MDL-59902 output: Added slasharguments to url of single_button
Since single_buttons used out_omit_querystring without specifying to
use slasharguments, it was not possible to use slasharguments with a
single button. Since slasharguments can also not be appended to the
form, which is sent by a single_button, this change was necessary to
enable usecases which need slasharguments within the button's url.